hpn #012 move setup/teardown into collection fixture
