+++ sources/samples/Eshva.Hodgepodge.LearningRedis/Eshva.Hodgepodge.LearningRedis.csproj
@@ -11,6 +11,7 @@
 
   <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="5.9.0" />
+    <PackageReference Include="JetBrains.Annotations" Version="2019.1.3" />
     <PackageReference Include="Microsoft.Extensions.FileProviders.Abstractions" Version="3.1.0" />
     <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="3.1.0" />
     <PackageReference Include="StackExchange.Redis" Version="2.0.601" />
